pyproject.toml
[project]
name = "calling-llm-apis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai==1.14.0",
    "grpcio==1.71.0",
    "ipdb==0.13.13",
    "ipykernel==6.29.5",
    "ipython==8.32.0",
    "ipywidgets==8.1.5",
    "jupyter-client==8.6.3",
    "jupyter-core==5.7.2",
    "jupyter-events==0.12.0",
    "jupyter-lsp==2.2.5",
    "jupyter-server==2.15.0",
    "jupyter-server-terminals==0.5.3",
    "jupyterlab==4.3.5",
    "jupyterlab-code-formatter==3.0.2",
    "jupyterlab-execute-time==3.2.0",
    "jupyterlab-pygments==0.3.0",
    "jupyterlab-server==2.27.3",
    "jupyterlab-widgets==3.0.13",
    "langchain==0.3.25",
    "langchain-chroma==0.2.3",
    "langchain-community==0.3.23",
    "langchain-core==0.3.59",
    "langchain-google-genai==2.1.4",
    "langchain-text-splitters==0.3.8",
    "langgraph==0.4.3",
    "langsmith==0.3.42",
    "nbclient==0.10.2",
    "nbconvert==7.16.6",
    "nbformat==5.10.4",
    "nbresult==0.1.1",
    "notebook==7.3.2",
    "notebook-shim==0.2.4",
    "numpy>2",
    "pandas==2.2.3",
    "protobuf<6",
    "pylint==3.3.4",
    "pypdf==5.4.0",
    "pytest==8.3.4",
    "python-dotenv==1.1.0",
    "wikipedia==1.4.0",
]
